refactor: 내부통신 client 패키지 변경
