+++ src/main/java/com/sejong/metaservice/application/common/config/FeignClientConfig.java
@@ -5,6 +5,6 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@EnableFeignClients(basePackages = "com.sejong.metaservice.infrastructure.client")
+@EnableFeignClients(basePackages = "com.sejong.metaservice.client")
 public class FeignClientConfig {
 }
