+++ src/main/java/com/sejong/metaservice/application/internal/CSKnowledgeInternalService.java
@@ -4,8 +4,8 @@ import static com.sejong.metaservice.core.common.exception.ExceptionType.BAD_REQ
 import static com.sejong.metaservice.core.common.exception.ExceptionType.EXTERNAL_SERVER_ERROR;
 
 import com.sejong.metaservice.application.internal.response.PostLikeCheckResponse;
+import com.sejong.metaservice.client.ArchiveClient;
 import com.sejong.metaservice.core.common.exception.BaseException;
-import com.sejong.metaservice.infrastructure.client.ArchiveClient;
 import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
